Reassign service data on POST with form data rather than updating
We were previously updating our brief response with data from
the form if there were user errors. Infact, we don't need to this
and we can just pass the form data to the template by itself. This
will fix the case where evidence data existed for a question and
then the user removes the text from the textbox and submits the form.
The empty string will be replayed to the user instead of data
which used to exist.