Hopefully make framework selection more intelligible
For the "Become a supplier" page we need to choose one framework for each "framework_framework" we know about to display a message about on the "Become a supplier" page.

If there's an "open" framework we want the message from that, if not then any "coming" framework is next on our list, and failing that a closed framework.

The logic for this was working with the previous version but has been described as "well, it throws around some lists and stuff, picks some values out and then something comes out the end".

Hopefully in this version it's clearer what the intent of the code is.